Retry everything in the integration tests
Kafka is really funky. You ask a broker for topic metadata, make sure the topic and partition you want to publish to exist, and then you ask the broker that is the leader for that topic and partition for the same metadata, to make sure it agrees, and then you publish, and you get UnknownTopicOrPartitionError back.